+++ src/main/java/edu/stevens/ssw690/DuckSource/dao/MailMessageDaoImpl.java
@@ -1,19 +1,22 @@
 package edu.stevens.ssw690.DuckSource.dao;
 
-import java.math.BigInteger;
 import java.util.List;
 
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
-import javax.persistence.Query;
 
 import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
-import edu.stevens.ssw690.DuckSource.model.DuckUser;
 import edu.stevens.ssw690.DuckSource.model.MailMessage;
 
+/**
+ * 
+ * @author susan
+ * @see MailMessageDao
+ *
+ */
 @Repository
 @Transactional
 @Component
@@ -37,7 +40,7 @@ public class MailMessageDaoImpl implements MailMessageDao {
 	 }
 	
 	
-	public MailMessage findById(BigInteger id) {
+	public MailMessage findById(Integer id) {
     	return em.find(MailMessage.class, id);
 	}
