+++ src/main/java/edu/stevens/ssw690/DuckSource/service/MailDistributionManager.java
@@ -1,16 +1,41 @@
 package edu.stevens.ssw690.DuckSource.service;
 
-import java.math.BigInteger;
 import java.util.List;
 
 import edu.stevens.ssw690.DuckSource.model.MailDistribution;
 
 public interface MailDistributionManager {
 	
+	/**
+	 * Saves the mail distribution to the database (managed) 
+	 * @param mailDistribution
+	 */
 	public void persist(MailDistribution mailDistribution);
+	
+	/**
+     * Saves the mail distribution to the database (unmanaged)
+     * @param mailDistribution
+     */
     public void merge(MailDistribution mailDistribution);
+    
+    /**
+     * Gets a list of all Mail Distributions
+     * @return List of Mail Distributions
+     */
 	public List<MailDistribution> getAll();
-	public List<MailDistribution> getByMessageId(BigInteger id);
-	public MailDistribution findById(BigInteger id);
+	
+	/**
+	 * Gets a list of all mail distribution foe a message by id
+	 * @param id
+	 * @return List of Mail Distributions
+	 */
+	public List<MailDistribution> getByMessageId(Integer id);
+	
+	/**
+	 * Get a message distribution by id
+	 * @param id
+	 * @return Mail Distribution
+	 */
+	public MailDistribution findById(Integer id);
 
 }
